mode connectors to web3-react module
